Fix cpuinfo dictionary keys
Fix raw_arch_string, brand, and vendor_id dictionary keys used
to access information returned by cpuinfo.get_cpu_info().

It also fixes a comparison regarding the cpuversion, which should
be against a tuple, not a string.

Finally, it adds a warning message when the device arch or type
can not be recognized (like when running on x86_64).

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>